django-api/pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["hvieira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10.13"
Django = "~5.1"
psycopg2 = "~2.9.9"
djangorestframework = "~3.15.2"

[tool.poetry.group.dev.dependencies]
pytest = "~8.3.2"
pytest-django = "~4.9.0"
coverage = "~7.6.1"
requests = "~2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
